moved integration tests to different folder
