+++ src/it/java/de/lh/tool/rest/InfoIT.java
@@ -1,4 +1,4 @@
-package de.lh.tool.integration;
+package de.lh.tool.rest;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -7,19 +7,19 @@ import java.io.IOException;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
-import de.lh.tool.TestUtil;
+import de.lh.tool.IntegrationTestUtil;
 import io.restassured.RestAssured;
 
 public class InfoIT {
 
 	@BeforeAll
 	public static void waitForServer() {
-		TestUtil.waitForLocalTomcat();
+		IntegrationTestUtil.waitForLocalTomcat();
 	}
 
 	@Test
 	public void testHeartbeat() throws IOException {
-		String url = TestUtil.REST_URL + "/info/heartbeat";
+		String url = IntegrationTestUtil.REST_URL + "/info/heartbeat";
 		assertTrue(Boolean.parseBoolean(RestAssured.get(url).asString()));
 	}
 }
