+++ src/test/java/de/lh/tool/config/security/JwtTokenProviderTest.java
@@ -1,4 +1,4 @@
-package de.lh.tool.unit.config.security;
+package de.lh.tool.config.security;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
